auto-install: large rewrite
Two things:
1) use sh instead of bash
2) transform monolithic script into separable functions

Using sh instead of bash will result in a faster running and simpler to
read and use script (plus, it wasn't leveraging many bashisms anyways).
By utilizing functions instead of just a stream of shell to execute, we
can add other features in a relatively straight-forward way. For
instance, adding signature or checksum verification to further ensure we
are installing the correct and intended snaps.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>